# backend/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"encoding/json"
	"time"

	"github.com/oapi-codegen/runtime"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for BusinessUserRole.
const (
	Business BusinessUserRole = "business"
)

// Defines values for CourierUserRole.
const (
	Courier CourierUserRole = "courier"
)

// Defines values for DeliveryStatus.
const (
	DeliveryStatusAccepted  DeliveryStatus = "accepted"
	DeliveryStatusDelivered DeliveryStatus = "delivered"
	DeliveryStatusPickedUp  DeliveryStatus = "picked_up"
	DeliveryStatusPosted    DeliveryStatus = "posted"
)

// Defines values for DeliveryPatchStatus.
const (
	DeliveryPatchStatusAccepted  DeliveryPatchStatus = "accepted"
	DeliveryPatchStatusDelivered DeliveryPatchStatus = "delivered"
	DeliveryPatchStatusPickedUp  DeliveryPatchStatus = "picked_up"
)

// Defines values for ListDeliveriesParamsStatus.
const (
	Accepted  ListDeliveriesParamsStatus = "accepted"
	Delivered ListDeliveriesParamsStatus = "delivered"
	PickedUp  ListDeliveriesParamsStatus = "picked_up"
	Posted    ListDeliveriesParamsStatus = "posted"
)

// BusinessUser defines model for BusinessUser.
type BusinessUser struct {
	BusinessAddress string           `firestore:"businessAddress"`
	BusinessName    string           `firestore:"businessName"`
	Email           string           `firestore:"email"`
	Id              string           `firestore:"id"`
	Location        GeoPoint         `firestore:"location"`
	PlaceId         *string          `firestore:"placeId,omitempty"`
	Role            BusinessUserRole `firestore:"role"`
}

// BusinessUserRole defines model for BusinessUser.Role.
type BusinessUserRole string

// CourierUser defines model for CourierUser.
type CourierUser struct {
	CourierName string          `firestore:"courierName"`
	Email       string          `firestore:"email"`
	Id          string          `firestore:"id"`
	Role        CourierUserRole `firestore:"role"`
}

// CourierUserRole defines model for CourierUser.Role.
type CourierUserRole string

// Delivery defines model for Delivery.
type Delivery struct {
	AssignedTo          *string        `firestore:"assignedTo"`
	BusinessAddress     string         `firestore:"businessAddress"`
	BusinessId          *string        `firestore:"businessId,omitempty"`
	BusinessLocation    GeoPoint       `firestore:"businessLocation"`
	BusinessName        string         `firestore:"businessName"`
	CreatedAt           *time.Time     `firestore:"createdAt,omitempty"`
	CreatedBy           *string        `firestore:"createdBy,omitempty"`
	DeliveredBy         *string        `firestore:"deliveredBy"`
	DestinationAddress  string         `firestore:"destinationAddress"`
	DestinationLocation GeoPoint       `firestore:"destinationLocation"`
	Id                  *string        `firestore:"id,omitempty"`
	Item                string         `firestore:"item"`
	Payment             float64        `firestore:"payment"`
	Status              DeliveryStatus `firestore:"status"`
}

// DeliveryStatus defines model for Delivery.Status.
type DeliveryStatus string

// DeliveryCreate defines model for DeliveryCreate.
type DeliveryCreate struct {
	BusinessAddress     string   `firestore:"businessAddress"`
	BusinessLocation    GeoPoint `firestore:"businessLocation"`
	BusinessName        string   `firestore:"businessName"`
	DestinationAddress  string   `firestore:"destinationAddress"`
	DestinationLocation GeoPoint `firestore:"destinationLocation"`
	Item                string   `firestore:"item"`
	Payment             float64  `firestore:"payment"`
}

// DeliveryPatch defines model for DeliveryPatch.
type DeliveryPatch struct {
	AssignedTo *string              `firestore:"assignedTo"`
	Status     *DeliveryPatchStatus `firestore:"status,omitempty"`
}

// DeliveryPatchStatus defines model for DeliveryPatch.Status.
type DeliveryPatchStatus string

// Error defines model for Error.
type Error struct {
	Message *string `firestore:"message,omitempty"`
}

// GeoPoint defines model for GeoPoint.
type GeoPoint struct {
	Lat float64 `firestore:"lat"`
	Lng float64 `firestore:"lng"`
}

// OneOfUser defines model for OneOfUser.
type OneOfUser struct {
	union json.RawMessage
}

// PageSize defines model for PageSize.
type PageSize = int

// PageToken defines model for PageToken.
type PageToken = string

// NotFound defines model for NotFound.
type NotFound = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// ListDeliveriesParams defines parameters for ListDeliveries.
type ListDeliveriesParams struct {
	Status *ListDeliveriesParamsStatus `form:"status,omitempty" firestore:"status,omitempty"`
	Lat    *float64                    `form:"lat,omitempty" firestore:"lat,omitempty"`
	Lng    *float64                    `form:"lng,omitempty" firestore:"lng,omitempty"`

	// R Radius in kilometres from (lat,lng)
	R         *float64   `form:"r,omitempty" firestore:"r,omitempty"`
	PageSize  *PageSize  `form:"pageSize,omitempty" firestore:"pageSize,omitempty"`
	PageToken *PageToken `form:"pageToken,omitempty" firestore:"pageToken,omitempty"`
}

// ListDeliveriesParamsStatus defines parameters for ListDeliveries.
type ListDeliveriesParamsStatus string

// CreateDeliveryJSONRequestBody defines body for CreateDelivery for application/json ContentType.
type CreateDeliveryJSONRequestBody = DeliveryCreate

// UpdateDeliveryJSONRequestBody defines body for UpdateDelivery for application/json ContentType.
type UpdateDeliveryJSONRequestBody = DeliveryPatch

// AsBusinessUser returns the union data inside the OneOfUser as a BusinessUser
func (t OneOfUser) AsBusinessUser() (BusinessUser, error) {
	var body BusinessUser
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromBusinessUser overwrites any union data inside the OneOfUser as the provided BusinessUser
func (t *OneOfUser) FromBusinessUser(v BusinessUser) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeBusinessUser performs a merge with any union data inside the OneOfUser, using the provided BusinessUser
func (t *OneOfUser) MergeBusinessUser(v BusinessUser) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

// AsCourierUser returns the union data inside the OneOfUser as a CourierUser
func (t OneOfUser) AsCourierUser() (CourierUser, error) {
	var body CourierUser
	err := json.Unmarshal(t.union, &body)
	return body, err
}

// FromCourierUser overwrites any union data inside the OneOfUser as the provided CourierUser
func (t *OneOfUser) FromCourierUser(v CourierUser) error {
	b, err := json.Marshal(v)
	t.union = b
	return err
}

// MergeCourierUser performs a merge with any union data inside the OneOfUser, using the provided CourierUser
func (t *OneOfUser) MergeCourierUser(v CourierUser) error {
	b, err := json.Marshal(v)
	if err != nil {
		return err
	}

	merged, err := runtime.JSONMerge(t.union, b)
	t.union = merged
	return err
}

func (t OneOfUser) MarshalJSON() ([]byte, error) {
	b, err := t.union.MarshalJSON()
	return b, err
}

func (t *OneOfUser) UnmarshalJSON(b []byte) error {
	err := t.union.UnmarshalJSON(b)
	return err
}